refactor(editor): block handle layout 분리
